WeWatchTheBees/backend/requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8,<4.0
django-bootstrap5>=23.0,<25.0

